+++ src/test/scala/edu/rit/cs/mmior/jsonoid/discovery/schemas/NumberSchemaSpec.scala
@@ -138,7 +138,10 @@ class NumberSchemaSpec extends UnitSpec {
   }
 
   it should "have no properties in the minimal property set" in {
-    NumberSchema(0.0)(PropertySets.MinProperties).properties shouldBe empty
+    NumberSchema(0.0)(
+      PropertySets.MinProperties,
+      JsonoidParams.defaultJsonoidParams
+    ).properties shouldBe empty
   }
 
   it should "show integers as a valid type" in {